Announce a balance change on the EverNifeCore bus
The 2.x EconomyUpdateEvent was a plain org.bukkit.event.Event, so it existed
only on Bukkit and only a Bukkit listener could hear it. On the bus it reaches
both platforms and stays audible to a plain @EventHandler, which is what lets
an economy bridge keep listening the way it always did.

It is posted with postIfListened, so with nobody subscribed a balance change
allocates nothing.

The balance itself becomes a BigDecimal: cents now add up exactly instead of
drifting through binary floating point, and Vault 2 already speaks that type.
The storage index cannot carry a decimal - IndexHint.FieldType stops at DOUBLE
- so a double projection of the balance carries the ordering while the exact
value stays in the payload.

Settings become one POJO per block, bound with getOrMergeValue. The compiler
now enforces the grouping the flat keys only suggested, and a derived value
like the baltop row cap or the placeholder cache floor lives next to the
setting it derives from instead of in every caller.

/eco stays the main command: CMDEco is now CMDFinalEconomy, carrying the
finaleconomy alias and the reload subcommand, and the reload-only class it
replaces is gone.

No class is final and no class hides its constructor - both were fences
against a use nobody had tried.

Gradle wrapper moves to 9.7.1.

+++ common/src/main/java/br/com/finalcraft/finaleconomy/common/economy/vault2/FinalEconomyVault2.java
@@ -53,7 +53,7 @@ public class FinalEconomyVault2 implements Economy {
 
     @Override
     public String format(BigDecimal amount) {
-        return EconomyService.format(amount.doubleValue());
+        return EconomyService.format(amount);
     }
 
     @Override
@@ -171,7 +171,7 @@ public class FinalEconomyVault2 implements Economy {
 
     @Override
     public BigDecimal getBalance(String pluginName, UUID accountID) {
-        return BigDecimal.valueOf(EconomyService.getBalance(accountID));
+        return EconomyService.getBalance(accountID);
     }
 
     @Override
@@ -186,7 +186,7 @@ public class FinalEconomyVault2 implements Economy {
 
     @Override
     public boolean has(String pluginName, UUID accountID, BigDecimal amount) {
-        return EconomyService.has(accountID, amount.doubleValue());
+        return EconomyService.has(accountID, amount);
     }
 
     @Override
@@ -201,7 +201,7 @@ public class FinalEconomyVault2 implements Economy {
 
     @Override
     public EconomyResponse withdraw(String pluginName, UUID accountID, BigDecimal amount) {
-        return responseOf(EconomyService.withdraw(accountID, amount.doubleValue()));
+        return responseOf(EconomyService.withdraw(accountID, amount));
     }
 
     @Override
@@ -216,7 +216,7 @@ public class FinalEconomyVault2 implements Economy {
 
     @Override
     public EconomyResponse deposit(String pluginName, UUID accountID, BigDecimal amount) {
-        return responseOf(EconomyService.deposit(accountID, amount.doubleValue()));
+        return responseOf(EconomyService.deposit(accountID, amount));
     }
 
     @Override
